Windows portability and build improvements
Get the current version number from configure.ac rather than configure
so that the Windows build scripts work from a Git checkout, and rename
configure.bat to configure.cmd.

Choose the correct GSS-API library for 64-bit Windows.

Quote the path to the Kerberos SDK on Windows.

Use SOCKET for the type of sockets rather than int, for correct handling
of 64-bit Windows.  Use INVALID_SOCKET rather than -1 for an invalid
socket value.

Disable warnings about unknown pragmata and POSIX strdup on Windows.

Cast the flag argument to setsockopt to avoid warnings on Windows.

Restructure the boundary logic in the xread function in util/tokens.c to
be more obvious for both humans and the compiler.

Cast the priority value passed into ReportEvent() appropriately.

+++ util/messages.c
@@ -182,7 +182,7 @@ message_log_syslog(int pri, int len, const char *fmt, va_list args, int err)
 
         eventlog = RegisterEventSource(NULL, message_program_name);
         if (eventlog != NULL) {
-            ReportEvent(eventlog, pri, 0, 0, NULL, 1, 0, &buffer, NULL);
+            ReportEvent(eventlog, (WORD)pri, 0, 0, NULL, 1, 0, &buffer, NULL);
             CloseEventLog(eventlog);
         }
     }
